+++ src/app.controller.spec.ts
@@ -1,7 +1,8 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { AppController } from './app.controller';
-import { AppService } from '../services/app.service';
 import { Chance } from 'chance';
+import { AppController } from './app.controller';
+import { AppService } from './app.service';
+
 const chance = new Chance();
 
 describe('AppController', () => {
