+++ src/app.module.ts
@@ -1,17 +1,17 @@
 import { GraphQLModule } from '@nestjs/graphql';
 import { Module } from '@nestjs/common';
-import { AppController } from './controllers/app.controller';
-import { AppService } from './services/app.service';
-import { AuthModule } from './resolvers/auth/auth.module';
-import { UserModule } from './resolvers/user/user.module';
-import { PostModule } from './resolvers/post/post.module';
-import { AppResolver } from './resolvers/app.resolver';
-import { DateScalar } from './common/scalars/date.scalar';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import config from './configs/config';
-import { GraphqlConfig } from './configs/config.interface';
 import { PrismaModule } from 'nestjs-prisma';
-import { loggingMiddleware } from './logging.middleware';
+import { AppController } from './app.controller';
+import { AppService } from './app.service';
+import { AppResolver } from './app.resolver';
+import { AuthModule } from 'src/auth/auth.module';
+import { UserModule } from 'src/users/user.module';
+import { PostModule } from 'src/posts/post.module';
+import { DateScalar } from 'src/common/scalars/date.scalar';
+import config from 'src/common/configs/config';
+import type { GraphqlConfig } from 'src/common/configs/config.interface';
+import { loggingMiddleware } from 'src/common/middleware/logging.middleware';
 
 @Module({
   imports: [
