+++ src/auth/auth.module.ts
@@ -1,13 +1,13 @@
-import { PasswordService } from './../../services/password.service';
-import { GqlAuthGuard } from '../../guards/gql-auth.guard';
-import { AuthService } from '../../services/auth.service';
-import { AuthResolver } from './auth.resolver';
 import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
-import { JwtStrategy } from './jwt.strategy';
 import { ConfigService } from '@nestjs/config';
-import { SecurityConfig } from 'src/configs/config.interface';
+import { PasswordService } from './password.service';
+import { GqlAuthGuard } from './gql-auth.guard';
+import { AuthService } from './auth.service';
+import { AuthResolver } from './auth.resolver';
+import { JwtStrategy } from './jwt.strategy';
+import { SecurityConfig } from 'src/common/configs/config.interface';
 
 @Module({
   imports: [
