+++ src/auth/auth.resolver.ts
@@ -1,6 +1,3 @@
-import { Auth } from '../../models/auth.model';
-import { Token } from '../../models/token.model';
-import { LoginInput } from './dto/login.input';
 import {
   Resolver,
   Mutation,
@@ -8,7 +5,10 @@ import {
   Parent,
   ResolveField,
 } from '@nestjs/graphql';
-import { AuthService } from '../../services/auth.service';
+import { AuthService } from './auth.service';
+import { Auth } from './models/auth.model';
+import { Token } from './models/token.model';
+import { LoginInput } from './dto/login.input';
 import { SignupInput } from './dto/signup.input';
 import { RefreshTokenInput } from './dto/refresh-token.input';
