+++ src/auth/jwt.strategy.ts
@@ -1,10 +1,10 @@
-import { JwtDto } from './dto/jwt.dto';
 import { Strategy, ExtractJwt } from 'passport-jwt';
 import { PassportStrategy } from '@nestjs/passport';
 import { Injectable, UnauthorizedException } from '@nestjs/common';
-import { AuthService } from '../../services/auth.service';
-import { User } from '@prisma/client';
 import { ConfigService } from '@nestjs/config';
+import { User } from '@prisma/client';
+import { AuthService } from './auth.service';
+import { JwtDto } from './dto/jwt.dto';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
