+++ src/common/pagination/pagination.ts
@@ -1,6 +1,6 @@
 import { Field, ObjectType, Int } from '@nestjs/graphql';
-import { PageInfo } from './page-info.model';
 import { Type } from '@nestjs/common';
+import { PageInfo } from './page-info.model';
 
 export default function Paginated<TItem>(TItemClass: Type<TItem>) {
   @ObjectType(`${TItemClass.name}Edge`)
