+++ src/posts/dto/createPost.input.ts
