+++ src/posts/models/post-connection.model.ts
@@ -1,6 +1,6 @@
 import { ObjectType } from '@nestjs/graphql';
-import PaginatedResponse from '../../common/pagination/pagination';
-import { Post } from '../post.model';
+import PaginatedResponse from 'src/common/pagination/pagination';
+import { Post } from './post.model';
 
 @ObjectType()
 export class PostConnection extends PaginatedResponse(Post) {}
