+++ src/posts/models/post.model.ts
@@ -1,6 +1,6 @@
 import { ObjectType } from '@nestjs/graphql';
-import { User } from './user.model';
-import { BaseModel } from './base.model';
+import { User } from 'src/users/models/user.model';
+import { BaseModel } from 'src/common/models/base.model';
 
 @ObjectType()
 export class Post extends BaseModel {
