+++ src/posts/post.module.ts
@@ -1,5 +1,5 @@
-import { PostResolver } from './post.resolver';
 import { Module } from '@nestjs/common';
+import { PostResolver } from './post.resolver';
 
 @Module({
   imports: [],
