+++ test/app.e2e-spec.ts
@@ -1,8 +1,9 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import * as request from 'supertest';
-import { AppModule } from '../src/app.module';
-import { Chance } from 'chance';
 import { INestApplication } from '@nestjs/common';
+import * as request from 'supertest';
+import { Chance } from 'chance';
+import { AppModule } from 'src/app.module';
+
 const chance = new Chance();
 
 describe('AppController (e2e)', () => {
